Renamed image files from micronaut* to grails*
The images are grails logos, the filenames should be grails*.png|.svg
