+++ app/launch/src/images/grails-white-icon.png
