+++ app/launch/src/images/grails-white.png
