+++ app/launch/src/images/grails.png
